refactor: move state to common/state
